+++ daiv/automation/agents/code_describer/agent.py
@@ -43,4 +43,4 @@ class CodeDescriberAgent(BaseAgent[Runnable[CodeSnippet, str]]):
         prompt = ChatPromptTemplate.from_messages([("system", system), ("human", human)])
         return (
             prompt.partial(language="Not specified") | self.get_model(model=settings.MODEL_NAME) | StrOutputParser()
-        ).with_config({"run_name": settings.NAME, "tags": [settings.NAME], "max_concurrency": 10})
+        ).with_config({"run_name": settings.NAME, "tags": [settings.NAME], "max_concurrency": 5})